Add from/until option to arxiv crawling
It seems that responses are not correctly ordered (e.g. the response for cursor '1054001' returns
supr-con/9609004 which was submitted in 1996). However, using date filters seems to return new
submissions correctly. This commit will switch to a 'date' mode if either a starting or ending date
is provided. In case of an ending date, it will look for the last known date fetched to allow crawl
resumption.